fix: refactoring + kdoc
